+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # CHANGELOG
 ## master
+## v0.1.1 (2015-09-22)
+- `Transformer` extends `AbstractTransformer`
+- Add `TransformerError`
+- Add `transform()` to create anonymous `Transformers` from functions (`transform`, `initialize`, `flush`)
+
 ## v0.1.0 (2015-09-16)
 - **Initial release**